seg: search for long keys

+++ seg.h
@@ -3,5 +3,5 @@ struct seg *seg_init(int n, long *beg, long *end);
 void *seg_nodeget(struct seg *seg, int node);
 void seg_nodeput(struct seg *seg, int node, void *data);
 int seg_nodeset(struct seg *seg, int node, int **set, int *cnt);
-void seg_search(struct seg *seg, int x, void (*cb)(int node));
+void seg_search(struct seg *seg, long x, void (*cb)(int node));
 void seg_free(struct seg *seg);
